don't set up a Logger if used as a module
we're using a CassBot of our own in a different channel. we can use
bot.py mostly unchanged by loading it in a .tac and changing around
defaults before starting it up, but its attempt to make a
TimedRotatingFileHandler in a non-existent directory (on import instead
of on run) is annoying to get around. this change would make it easier.